fix: split routing constants to i18n/routing.ts and make redirect no-op on missing meta
- extract Locale, locales, defaultLocale, getLocalizedPath into src/i18n/routing.ts
  so client-side redirect code imports only routing constants, not translation objects
- re-export all symbols from src/i18n/index.ts to keep existing consumers unchanged
- update locale-redirect.ts to import from routing.ts instead of index.ts
- replace throw in NeutralLocaleRedirectPage.astro with graceful return so the page
  remains usable via manual locale links if the meta tag is absent

Resolves Copilot review comments on PR #59.
